DailyReportsListのカード部分を関数型コンポーネントに分離
