+++ src/components/pages/DailyReport/DailyReportsList.vue
@@ -54,30 +54,12 @@
             :key="post.uuid"
             :cols="(12 / itemsPerRow)"
           >
-            <v-card>
-              <v-card-subtitle class="pb-0">
-                {{ dateFormatter(post.created_at) }}
-              </v-card-subtitle>
-              <v-card-title>
-                {{ `#${post.id} ${post.title}` }}
-              </v-card-title>
-              <v-card-text
-               class="text--primary text-truncate d-inline-block"
-               style="max-width: 180px"
-              >
-                {{ removeHtmlTag(post.body_text) }}
-              </v-card-text>
-              <v-card-actions>
-                <v-btn
-                  text
-                  color="blue-grey lighten-1"
-                  link
-                  :to="`/daily_reports/post/${post.id}`"
-                >
-                  {{ $t('dailyReport.readMore') }}
-                </v-btn>
-              </v-card-actions>
-            </v-card>
+            <DailyReportsCard
+              :id="post.id"
+              :title="post.title"
+              :body_text="post.body_text"
+              :created_at="post.created_at"
+            />
           </v-col>
         </v-row>
       </template>
@@ -127,9 +109,12 @@
 </template>
 
 <script>
-import i18next from 'i18next'
+import DailyReportsCard from '@/components/modules/DailyReports/DailyReportsCard'
 export default {
   name: 'daily-reports-list',
+  components: {
+    DailyReportsCard
+  },
   props: ['isTopPage'],
   data () {
     return {
@@ -165,26 +150,6 @@ export default {
         .catch(_ => {
           this.reportPosts = []
         })
-    },
-    dateFormatter (date) {
-      var datetime = new Date(`${date}Z`)
-      var yearString = datetime.getFullYear()
-      var monthString = `0${datetime.getMonth() + 1}`.slice(-2) // 0埋め
-      var dateString = `0${datetime.getDate()}`.slice(-2)
-      var dayString = i18next.t(`commons.dayNameShort.${datetime.getDay()}`)
-      var hourString = `0${datetime.getHours()}`.slice(-2)
-      var minuteString = `0${datetime.getMinutes()}`.slice(-2)
-      var secondsString = `0${datetime.getSeconds()}`.slice(-2)
-      var formattedTime = `${yearString}/${monthString}/${dateString} (${dayString}) ${hourString}:${minuteString}:${secondsString}`
-
-      return formattedTime
-    },
-    removeHtmlTag (txt) {
-      if (txt != null && txt.match(/<("[^"]*"|'[^']*'|[^'">])*>/g)) {
-        return txt.replace(/<("[^"]*"|'[^']*'|[^'">])*>/g, '')
-      } else {
-        return txt
-      }
     },
     nextPage () {
       if (this.page + 1 <= this.numberOfPages) this.page += 1
